feat: update configuration template file and add dircommand for run
